chore: pull sn_client out of the node codebase
