+++ sn/benches/upload_bytes.rs
@@ -8,8 +8,9 @@
 
 use criterion::{criterion_group, criterion_main, Criterion};
 use eyre::Result;
-use safe_network::client::{
-    utils::test_utils::read_network_conn_info, Client, ClientConfig, Error,
+use sn_client::{
+    utils::test_utils::read_network_conn_info,
+    Client, ClientConfig, Error,
 };
 use sn_interface::types::utils::random_bytes;
 use sn_interface::types::Scope;
