+++ sn/examples/client_files.rs
@@ -7,7 +7,7 @@
 // permissions and limitations relating to use of the SAFE Network Software.
 
 use eyre::Result;
-use safe_network::client::{utils::test_utils::read_network_conn_info, Client, ClientConfig};
+use sn_client::{utils::test_utils::read_network_conn_info, Client, ClientConfig};
 use sn_interface::types::{utils::random_bytes, Scope};
 use std::time::Duration;
 use tokio::time::sleep;
