--- sn/src/client/codeowners
@@ -1,1 +1,0 @@
-* @maidsafe/backend_codeowners
