+++ sn/src/node/core/messaging/handling/join.rs
@@ -6,12 +6,12 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::elder_count;
 use crate::node::{
     api::cmds::Cmd,
     core::{relocation::RelocateDetailsUtils, Node},
     Result,
 };
+use sn_interface::elder_count;
 use sn_interface::messaging::system::{
     JoinAsRelocatedRequest, JoinAsRelocatedResponse, JoinRejectionReason, JoinRequest,
     JoinResponse, MembershipState, NodeState, SystemMsg,
