+++ sn/src/node/core/messaging/handling/relocation.rs
@@ -6,7 +6,6 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::elder_count;
 use crate::node::{
     api::cmds::Cmd,
     core::{
@@ -16,6 +15,7 @@ use crate::node::{
     },
     Event, Result,
 };
+use sn_interface::elder_count;
 use sn_interface::messaging::system::{
     MembershipState, NodeState as NodeStateMsg, RelocateDetails, SectionAuth,
 };
