+++ sn_api/Cargo.toml
@@ -34,8 +34,8 @@ rmp-serde = "1.0.0"
 serde = "1.0.123"
 serde_json = "1.0.62"
 sha3 = "~0.9"
-safe_network = { path = "../sn", version = "^0.58.13" }
-sn_interface = { path = "../sn_interface", version = "^0.1.1" }
+sn_client = { path = "../sn_client", version = "^0.1.0" }
+sn_interface = { path = "../sn_interface", version = "^0.1.0" }
 thiserror = "1.0.23"
 time = { version = "~0.3.4", features = ["formatting"] }
 tracing = "~0.1.26"
