+++ sn_api/src/app/files/mod.rs
@@ -22,7 +22,7 @@ use file_system::{
 use files_map::add_or_update_file_item;
 use log::{debug, info, warn};
 use relative_path::RelativePath;
-use safe_network::client::Client;
+use sn_client::Client;
 use sn_interface::types::BytesAddress;
 use std::{
     collections::{BTreeMap, HashSet},
