+++ sn_api/src/app/mod.rs
@@ -36,7 +36,7 @@ use super::{common, constants, Error, Result};
 use crate::NodeConfig;
 
 use rand::rngs::OsRng;
-use safe_network::client::{Client, ClientConfig, DEFAULT_OPERATION_TIMEOUT};
+use sn_client::{Client, ClientConfig, DEFAULT_OPERATION_TIMEOUT};
 use sn_interface::types::Keypair;
 use tracing::debug;
