+++ sn_client/src/api/data/pac_man.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::client::{Error, Result};
+use crate::{Error, Result};
 use bincode::serialize;
 use bytes::Bytes;
 use rayon::prelude::*;
