+++ sn_client/src/api/mod.rs
@@ -14,13 +14,13 @@ mod register_apis;
 
 pub use register_apis::RegisterWriteAheadLog;
 
-use crate::client::{connections::Session, errors::Error, ClientConfig};
-use crate::utils::read_prefix_map_from_disk;
+use crate::{connections::Session, errors::Error, ClientConfig};
 use sn_interface::messaging::{
     data::{CmdError, DataQuery, RegisterQuery, ServiceMsg},
     ServiceAuth, WireMsg,
 };
 use sn_interface::network_knowledge::prefix_map::NetworkPrefixMap;
+use sn_interface::network_knowledge::utils::read_prefix_map_from_disk;
 use sn_interface::types::{Chunk, Keypair, Peer, PublicKey, RegisterAddress};
 
 use bytes::Bytes;
@@ -281,9 +281,7 @@ impl Client {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::client::utils::test_utils::{
-        create_test_client, create_test_client_with, init_test_logger,
-    };
+    use crate::utils::test_utils::{create_test_client, create_test_client_with, init_test_logger};
     use eyre::Result;
     use sn_interface::types::utils::random_bytes;
     use sn_interface::types::Scope;
