+++ sn_client/src/connections/messaging.rs
@@ -8,14 +8,14 @@
 
 use super::{QueryResult, Session};
 
-use crate::client::{connections::CmdResponse, Error, Result};
-use crate::{at_least_one_correct_elder, elder_count};
+use crate::{connections::CmdResponse, Error, Result};
 use sn_interface::messaging::{
     data::{CmdError, DataQuery, QueryResponse},
     DstLocation, MsgId, MsgKind, ServiceAuth, WireMsg,
 };
 use sn_interface::network_knowledge::prefix_map::NetworkPrefixMap;
 use sn_interface::types::{Peer, PeerLinks, PublicKey, SendToOneError};
+use sn_interface::{at_least_one_correct_elder, elder_count};
 
 use backoff::{backoff::Backoff, ExponentialBackoff};
 use bytes::Bytes;
