+++ sn_client/src/errors.rs
