+++ sn_client/src/sample_config/sn_client.default.config
